Added traductor for sync logic

+++ Modules/Configuracion/Resolvers/SyncConfiguracionResolver.php
@@ -9,16 +9,28 @@
 namespace Modules\Configuracion\Resolvers;
 
 
+use Modules\Common\Resolvers\GenerateCodeResolverInterface;
+use Modules\Configuracion\Entities\Configuracion;
 use Modules\Configuracion\Repositories\ConfiguracionRepository;
 use Modules\Sincronizacion\Entities\Syncronizacion;
 
 class SyncConfiguracionResolver implements SyncConfiguracionResolverInterface
 {
+    /**
+     * @var ConfiguracionRepository
+     */
     private $configuracionRepository;
 
-    public function __construct(ConfiguracionRepository $configuracionRepository)
+    /**
+     * @var GenerateCodeResolverInterface
+     */
+    private $generateCodeResolver;
+
+    public function __construct(ConfiguracionRepository $configuracionRepository,
+                                GenerateCodeResolverInterface $generateCodeResolver)
     {
         $this->configuracionRepository = $configuracionRepository;
+        $this->generateCodeResolver = $generateCodeResolver;
     }
 
     public function handle(Syncronizacion $sincronizacion)
@@ -28,6 +40,15 @@ class SyncConfiguracionResolver implements SyncConfiguracionResolverInterface
             case 'INSERT':
                 $data = json_decode($sincronizacion->data,true);
                 $data['user_id'] = $sincronizacion->user_id;
+                $code = $data['clave'];
+
+                $validateCode = $this->configuracionRepository->validateCode($code);
+
+                if ($validateCode)
+                    $code = $this->generateCodeResolver->handle($code, $sincronizacion->tabla);
+
+                $data['clave'] = $code;
+
                 $this->configuracionRepository->create($data);
                 break;
             case 'UPDATE':
@@ -35,6 +56,7 @@ class SyncConfiguracionResolver implements SyncConfiguracionResolverInterface
                 $configuracion = $this->configuracionRepository->all()
                     ->where('clave', '=', $data['clave'])
                     ->first();
+                if ($configuracion)
                 $this->configuracionRepository->update($data, $configuracion->id);
                 break;
             case 'DELETE':
@@ -42,6 +64,7 @@ class SyncConfiguracionResolver implements SyncConfiguracionResolverInterface
                 $configuracion = $this->configuracionRepository->all()
                     ->where('clave', '=', $data['clave'])
                     ->first();
+                if ($configuracion)
                 $this->configuracionRepository->delete($configuracion->id);
                 break;
         }
